Move camera setting functions to setting class
